+++ lib/generators/foxinator/setup/templates/roles_and_permissions/app/controllers/application_controller.rb
@@ -10,7 +10,7 @@ class ApplicationController < ActionController::Base
   #
   #
   
-  include Routify, Sortify, Localizify, BackRedirectify
+  include Routify, Sortify, Localizify, Redirectify
   
   #
   # Settings