+++ lib/generators/foxinator/setup/templates/roles_and_permissions/app/controllers/concerns/redirectify.rb
@@ -1,4 +1,4 @@
-module BackRedirectify
+module Redirectify
   extend ActiveSupport::Concern
 
   included do