+++ changelog.md
@@ -1,4 +1,5 @@
 Dates are in American format.
 
+- 07/23/2015: Renamed BackRedirectify to Redirectify, added a tooltip helper (see README for instructions)
 - 07/20/2015: Added BackRedirectify concern for roles & permissions
 - 07/17/2015: Added breadcrumbs